% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{llpFixedControl}
\alias{llpFixedControl}
\title{Control options for fixed-value likelihood profiling}
\usage{
llpFixedControl()
}
\value{
A validated list of control options for fixed-value likelihood
profiling
}
\description{
Control options for fixed-value likelihood profiling
}
\seealso{
\code{\link[=llpProfileFixed]{llpProfileFixed()}}

Other Profiling: 
\code{\link{llpProfileFixed}()},
\code{\link{profile.nlmixr2FitCore}()},
\code{\link{profileNlmixr2FitCoreRet}()},
\code{\link{runLLP}()},
\code{\link{runLLPControl}()}
}
\concept{Profiling}
