% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{llpProfileFixed}
\alias{llpProfileFixed}
\alias{llpProfileFixedSingle}
\title{Estimate the objective function values for a model while fixing defined
parameter values}
\usage{
llpProfileFixed(fitted, which, control = list())

llpProfileFixedSingle(fitted, which)
}
\arguments{
\item{fitted}{The fit model}

\item{which}{A data.frame with column names of parameters to fix and values
of the fitted value to fix (one row only).}

\item{control}{A list passed to \code{llpFixedControl()} (currently unused)}
}
\value{
\code{which} with a column named \code{OFV} added with the objective function
value of the fitted estimate fixing the parameters in the other columns
}
\description{
Estimate the objective function values for a model while fixing defined
parameter values
}
\section{Functions}{
\itemize{
\item \code{llpProfileFixedSingle()}: Estimate the objective function value for a model
while fixing a single set of defined parameter values (for use in parameter
profiling)

}}
\seealso{
Other Profiling: 
\code{\link{llpFixedControl}()},
\code{\link{profile.nlmixr2FitCore}()},
\code{\link{profileNlmixr2FitCoreRet}()},
\code{\link{runLLP}()},
\code{\link{runLLPControl}()}

Other Profiling: 
\code{\link{llpFixedControl}()},
\code{\link{profile.nlmixr2FitCore}()},
\code{\link{profileNlmixr2FitCoreRet}()},
\code{\link{runLLP}()},
\code{\link{runLLPControl}()}
}
\author{
Bill Denney (changed by Matt Fidler to take out R 4.1 specific code)
}
\concept{Profiling}
